requirements.txt
tensorflow
keras
Pillow
pyserial
cv2
numpy
pyserial
